Use weak symbols
This is to allow the application to override the auto library
initialization and allow it to roll its own, should it prefer to do it
that way. One reason to do this is to enforce a certain minimal
version as the auto open library variants currently do not demand any
minimal version at all.